Add files via upload
TP v0.3
Agregue 2 opciones en el switch de modificar para detectar 's' y 'N'
Agregue la opcion de dar de baja un chofer.

+++ Trabajo Practico Empresa de transporte/Trabajo Practico Empresa de transporte/rlutil.h
@@ -574,7 +574,7 @@ RLUTIL_INLINE void locate(int x, int y) {
 #ifdef __cplusplus
 RLUTIL_INLINE void setString(const RLUTIL_STRING_T & str_) {
 	const char * const str = str_.data();
-	unsigned int len = str_.size();
+	size_t len = str_.size();
 #else // __cplusplus
 RLUTIL_INLINE void setString(RLUTIL_STRING_T str) {
 	unsigned int len = strlen(str);
